Replacing offset and limit from long to int

+++ app/src/main/java/br/com/kosawalabs/apprecommendation/data/AppDataRepository.java
@@ -8,11 +8,11 @@ import br.com.kosawalabs.apprecommendation.data.pojo.PackageName;
 public interface AppDataRepository {
     void setToken(String token);
 
-    void getApps(Long offset, Long limit, DataCallback<List<App>> callback);
+    void getApps(Integer offset, Integer limit, DataCallback<List<App>> callback);
 
     void getApp(Integer appId, DataCallback<App> callback);
 
-    void getRecommendedApps(Long offset, Long limit, DataCallback<List<App>> callback);
+    void getRecommendedApps(Integer offset, Integer limit, DataCallback<List<App>> callback);
 
     void sendMyApps(List<PackageName> packageNames, DataCallback<Void> callback);
 }
